Replace the old file by the CLI script
