+++ PhpRbac/install.php
@@ -1,322 +1,181 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-        <title>PHP-RBAC Installation</title>
-
-        <link rel="stylesheet" href="http://yui.yahooapis.com/pure/0.4.1/pure-min.css">
-
-    </head>
-
-    <body>
-
-        <div class="pure-g-r">
-            <div class="pure-u-1-3"></div>
-
-            <div class="pure-u-1-3">
-
-<?php if ((!isset($_POST['submit'])) && (!isset($_GET['process'])) && (!isset($_GET['success']))): ?>
-
 <?php
 
-if (isset($_GET['errorsExist'])) {
-
-    if (isset($_GET['blankFields'])) {
-        $fieldsRequired = true;
+    use PhpRbac\Rbac;
+    /**
+     * We check if the script is really executed by a shell user
+     * If he doesn't, we redirect the user to the website root
+     */
+    if(!isset($argv))
+    {
+        header('Location: /');
     }
-
-    if (isset($_GET['error_pw_match'])) {
-        $error_pw_match = true;
-    }
-
-    if (isset($_GET['error_dbAdapter'])) {
-        if ($_GET['error_dbAdapter'] == 1) {
-            $dbAdapter = '';
-        } else {
-            $dbAdapter = $_GET['error_dbAdapter'];
+    
+    require('autoload.php');
+    
+    $database = [
+        'adapter'       => null,
+        'host'          => null,
+        'name'          => null,
+        'table_prefix'  => null,
+        'user'          => null,
+        'password'      => null
+    ];
+    
+    $adapters = [
+        'pdo_mysql' => '"%s"',
+        'mysqli' => '"%s"',
+        'pdo_sqlite' => '__DIR__ . "/%s.sqlite3"' 
+    ];
+    
+    do{
+        $retry = true;
+        
+        setAdapters($database, $adapters);
+        setHost($database);
+        setName($database);
+        setTablePrefix($database);
+        setUser($database);
+        setPassword($database);
+        
+        if(confirm($database) === false)
+        {
+            fputs(STDOUT, 'Cleaning the previous data..' . PHP_EOL);
+            continue;
+        }
+        
+        generateConfig($database, $adapters);
+        generateDatabase();
+        
+        if(tryConnection($database) === false)
+        {
+            $retry = true;
+            continue;
+        }
+        
+        $retry = false;
+    }while($retry === true);
+    
+    function setAdapters(&$database, $adapters)
+    {
+        fputs(STDOUT, 'Please select one DB adapter' . PHP_EOL);
+        fputs(STDOUT, 'Available adapters' . PHP_EOL);
+        
+        reset($adapters);
+        while($key = key($adapters))
+        {
+            fputs(STDOUT, "- $key".PHP_EOL);
+            next($adapters);
+        }
+        
+        fputs(STDOUT, 'Database adapter [pdo_mysql] : ');
+        $adapter = trim(fgets(STDIN));
+        
+        $database['adapter'] = (!empty($adapter)) ? $adapter : 'pdo_mysql';
+        
+        if(!isset($adapters[$database['adapter']]))
+        {
+            fputs(STDOUT, 'A wrong adapter has been chosen. Please correct it.' . PHP_EOL);
+            setAdapters($database, $adapters);
+            return true;
         }
     }
+    
+    function setHost(&$database)
+    {
+        fputs(STDOUT, 'Database host (ex. 127.0.0.1) : ');
 
-    if (isset($_GET['error_dbHost'])) {
-        if ($_GET['error_dbHost'] == 1) {
-            $dbHost = '';
-        } else {
-            $dbHost = $_GET['error_dbHost'];
+        $database['host'] = trim(fgets(STDIN));
+    }
+    
+    function setName(&$database)
+    {
+        fputs(STDOUT, 'Database name : ');
+
+        $database['name'] = trim(fgets(STDIN));
+    }
+    
+    function setTablePrefix(&$database)
+    {
+        fputs(STDOUT, 'Table prefix [phprbac_] : ');
+        
+        $prefix = trim(fgets(STDIN));
+        $database['table_prefix'] = (!empty($prefix)) ? $prefix : 'phprbac_';
+    }
+    
+    function setUser(&$database)
+    {
+        fputs(STDOUT, 'Database user : ');
+
+        $database['user'] = trim(fgets(STDIN));
+    }
+    
+    function setPassword(&$database)
+    {
+        fputs(STDOUT, 'Database password : ');
+
+        $database['password'] = trim(fgets(STDIN));
+    }
+    
+    function confirm($database)
+    {
+        fputs(STDOUT, PHP_EOL);
+        fputs(STDOUT, "The host is {$database['host']}" . PHP_EOL);
+        fputs(STDOUT, "The database name is {$database['name']} with {$database['adapter']} driver" . PHP_EOL);
+        fputs(STDOUT, "All the RBAC tables will be prefixed with {$database['table_prefix']}" . PHP_EOL);
+        fputs(STDOUT, "The credentials are {$database['user']}:{$database['password']}" . PHP_EOL);
+        fputs(STDOUT, "Do you confirm this is the expected configuration [Y/n] (Y) ?" . PHP_EOL);
+        
+        return (trim(fgets(STDIN)) !== 'n');
+    }
+    
+    function tryConnection($database)
+    {
+        try{
+            new PDO(
+                "mysql:host={$database['host']}; dbname={$database['name']}",
+                $database['user'],
+                $database['password'],
+                [
+                    PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION
+                ]
+            );
+            fputs(STDOUT, 'Test connection success.' . PHP_EOL);
+            return true;
+        }
+        catch (PDOException $ex)
+        {
+            fputs(STDOUT, 'The connection failed. Please try again.' . PHP_EOL);
+            return false;
         }
     }
+    
+    function generateConfig($database, $adapters)
+    {
+        $data =	 
+            '<?php' . PHP_EOL .
+            '$adapter="' . $database['adapter'] . '";' . PHP_EOL . 
+            '$host="' . $database['host'] . '";'. PHP_EOL .
+            '$dbname=' . sprintf($adapters[$database['adapter']], $database['name']) . ';' . PHP_EOL .
+            '$tablePrefix = "' . $database['table_prefix'] . '";' . PHP_EOL .
+            '$user="' . $database['user'] . '";' . PHP_EOL .
+            '$pass="' . $database['password'] . '";' . PHP_EOL
+        ;
 
-    if (isset($_GET['error_dbName'])) {
-        if ($_GET['error_dbName'] == 1) {
-            $dbName = '';
-        } else {
-            $dbName = $_GET['error_dbName'];
+        $dbConnFile = 'database' . DIRECTORY_SEPARATOR . 'database.config';
+
+        file_put_contents($dbConnFile, $data);
+
+        $currentOS = strtoupper(substr(PHP_OS, 0, 3));
+
+        if ($currentOS != 'WIN') {
+            chmod($dbConnFile, 0644);
         }
     }
-
-    if (isset($_GET['error_dbTablePrefix'])) {
-        if ($_GET['error_dbTablePrefix'] == 1) {
-            $dbTablePrefix = '';
-        } else {
-            $dbTablePrefix = $_GET['error_dbTablePrefix'];
-        }
+    
+    function generateDatabase()
+    {
+        $rbac = new Rbac();
+        $rbac->reset(true);
+        
+        fputs(STDOUT, 'Database is generated.' . PHP_EOL);
     }
-
-    if (isset($_GET['error_dbUser'])) {
-        if ($_GET['error_dbUser'] == 1) {
-            $dbUser = '';
-        } else {
-            $dbUser = $_GET['error_dbUser'];
-        }
-    }
-
-} else {
-    $dbAdapter = 'MySQL';
-    $dbHost = '';
-    $dbName = '';
-    $dbTablePrefix = '';
-    $dbUser = '';
-}
-
-?>
-
-<h1>PHP-RBAC Quick Install</h1>
-
-<h2>Database information</h2>
-
-<?php if (isset($fieldsRequired)): ?>
-
-<p>
-    <strong style="color: red">All Fields Are Required</strong>
-</p>
-
-<?php elseif (isset($_GET['error_pw_match'])) :?>
-
-<p>
-    <strong style="color: red">Passwords Do Not Match</strong>
-</p>
-
-<?php endif; ?>
-
-<form class="pure-form pure-form-aligned" name="dbInfo" action="" method="POST">
-
-    <fieldset>
-
-        <div class="pure-control-group">
-            <label for="dbAdapter">Database Adapter:</label>
-            <select id="dbAdapter" name="dbAdapter">
-
-                    <?php if ($dbAdapter == 'MySQL'): ?>
-                        <option selected="selected">MySQL</option>
-                    <?php else: ?>
-                        <option>MySQL</option>
-                    <?php endif; ?>
-
-                    <?php if ($dbAdapter == 'MySQLi'): ?>
-                        <option selected="selected">MySQLi</option>
-                    <?php else: ?>
-                        <option>MySQLi</option>
-                    <?php endif; ?>
-
-                    <?php if ($dbAdapter == 'SQLite'): ?>
-                        <option selected="selected">SQLite</option>
-                    <?php else: ?>
-                        <option>SQLite</option>
-                    <?php endif; ?>
-
-                </select>
-        </div>
-
-        <div class="pure-control-group">
-            <label for="dbHost">Database Host Name/IP:</label>
-            <input id="dbHost" type="text" name="dbHost" value="<?php echo $dbHost; ?>" required />
-        </div>
-
-        <div class="pure-control-group">
-            <label for="dbName">Database Name:</label>
-            <input id="dbName" type="text" name="dbName" value="<?php echo $dbName; ?>" required />
-        </div>
-
-        <div class="pure-control-group">
-            <label for="dbTablePrefix">Table Prefix:</label>
-            <input id="dbTablePrefix" type="text" name="dbTablePrefix" value="<?php echo ($dbTablePrefix == '') ? 'phprbac_' : $dbTablePrefix;  ?>" required />
-        </div>
-
-        <div class="pure-control-group">
-            <label for="dbUser">Database User:</label>
-            <input id="dbUser" type="text" name="dbUser" value="<?php echo $dbUser; ?>" required />
-        </div>
-
-        <div class="pure-control-group">
-            <label for="dbPassword">Password:</label>
-            <input id="dbPassword" type="password" name="dbPassword" value="" required />
-        </div>
-
-        <div class="pure-control-group">
-            <label for="dbPasswordConfirm">Confirm Password:</label>
-            <input id="dbPasswordConfirm" type="password" name="dbPasswordConfirm" value="" required />
-        </div>
-
-        <div class="pure-controls">
-            <button type="submit" name="submit" class="pure-button pure-button-primary">Submit</button>
-        </div>
-
-    </fieldset>
-
-</form>
-
-<?php elseif(isset($_POST['submit'])): ?>
-
-    <?php
-
-    // Set error flag
-    $errorExists = false;
-
-    // Start error string
-    $errors = '?errorsExist=1&blankFields=1&error_dbAdapter=' . $_POST['dbAdapter'] . '&';
-
-    // Make sure there are no empty fields
-    if (empty($_POST['dbHost'])) {
-        // Create error
-        $errorExists = true;
-        $errors .= 'error_dbHost=1&';
-    } else {
-        $errors .= 'error_dbHost=' . $_POST['dbHost'] . '&';
-    }
-
-    if (empty($_POST['dbName'])) {
-        // Create error
-        $errorExists = true;
-        $errors .= 'error_dbName=1&';
-    } else {
-        $errors .= 'error_dbName=' . $_POST['dbName'] . '&';
-    }
-
-    if (empty($_POST['dbTablePrefix'])) {
-        // Create error
-        $errorExists = true;
-        $errors .= 'error_dbTablePrefix=1&';
-    } else {
-        $errors .= 'error_dbTablePrefix=' . $_POST['dbTablePrefix'] . '&';
-    }
-
-    if (empty($_POST['dbUser'])) {
-        // Create error
-        $errorExists = true;
-        $errors .= 'error_dbUser=1&';
-    } else {
-        $errors .= 'error_dbUser=' . $_POST['dbUser'] . '&';
-    }
-
-    if (empty($_POST['dbPassword'])) {
-        // Create error
-        $errorExists = true;
-        $errors .= 'error_dbPassword=1&';
-    } else {
-        $errors .= 'error_dbPassword=' . $_POST['dbPassword'] . '&';
-    }
-
-    if (empty($_POST['dbPasswordConfirm'])) {
-        // Create error
-        $errorExists = true;
-        $errors .= 'error_dbPasswordConfirm=1&';
-    } else {
-        $errors .= 'error_dbPasswordConfirm=' . $_POST['dbPasswordConfirm'] . '&';
-    }
-
-    $errors = rtrim($errors, '&');
-
-    if ($errorExists === true) {
-        header('Location: install.php' . $errors);
-        exit;
-    }
-
-    // Make sure the passwords match
-
-    if ($_POST['dbPassword'] === $_POST['dbPasswordConfirm']) {
-        // Pass
-        header('Location: install.php?process=1&dbAdapter=' . $_POST['dbAdapter'] . '&dbHost=' . $_POST['dbHost'] . '&dbName=' . $_POST['dbName'] . '&dbTablePrefix=' . $_POST['dbTablePrefix'] . '&dbUser=' . $_POST['dbUser'] . '&dbPassword=' . $_POST['dbPassword'] . '&dbPasswordConfirm=' . $_POST['dbPasswordConfirm']);
-        exit;
-    } else {
-        // Fail
-        header('Location: install.php?errorsExist=1&error_pw_match=1&error_dbAdapter=' . $_POST['dbAdapter'] . '&error_dbHost=' . $_POST['dbHost'] . '&error_dbName=' . $_POST['dbName'] . '&error_dbTablePrefix=' . $_POST['dbTablePrefix'] . '&error_dbUser=' . $_POST['dbUser']);
-        exit;
-    }
-
-    ?>
-
-<?php elseif(isset($_GET['process'])): ?>
-
-    <?php
-
-    // Create database.config
-
-    if ($_GET['dbAdapter'] == 'MySQL') {
-        $dbAdapterString = '$adapter="pdo_mysql";';
-        $dbNameString = '$dbname="' . $_GET['dbName'] . '";';
-    } elseif ($_GET['dbAdapter'] == 'MySQLi') {
-        $dbAdapterString = '$adapter="mysqli";';
-        $dbNameString = '$dbname="' . $_GET['dbName'] . '";';
-    } else {
-        $dbAdapterString = '$adapter="pdo_sqlite";';
-        $dbNameString = '$dbname=__DIR__."/' . $_GET['dbName'] . '.sqlite3";';
-    }
-
-    $data =	 '<?php
-
-' . $dbAdapterString . '
-$host="' . $_GET['dbHost'] . '";
-
-' . $dbNameString . '
-$tablePrefix = "' . $_GET['dbTablePrefix'] . '";
-
-$user="' . $_GET['dbUser'] . '";
-$pass="' . $_GET['dbPassword'] . '";
-
-';
-
-    $dbConnFile = 'database' . DIRECTORY_SEPARATOR . 'database.config';
-
-    file_put_contents($dbConnFile, $data);
-
-    $currentOS = strtoupper(substr(PHP_OS, 0, 3));
-
-    if ($currentOS != 'WIN') {
-        chmod($dbConnFile, 0644);
-    }
-
-    // Instantiate PhpRbac\Rbac object
-    require_once 'autoload.php';
-
-    $rbac = new PhpRbac\Rbac();
-
-    // execute '$rbac->reset(true);'
-    $rbac->reset(true);
-
-    // Send to Success Message
-    header('Location: install.php?success=1');
-    exit;
-
-    ?>
-
-<?php elseif(isset($_GET['success'])): ?>
-
-    <h1>Installation Successful!</h1>
-
-    <h2>Congratulations: You are now ready to use PHP-RBAC!</h2>
-
-    <p style="color: red">
-        <strong>Warning - Please remove 'install.php' for security reasons!</strong>
-    </p>
-
-<?php endif; ?>
-
-            </div>
-
-            <div class="pure-u-1-3"></div>
-
-        </div>
-
-    </body>
-</html>
